requirements.txt
Pyrogram==2.0.106
aiogram==3.10.0
PyYAML==6.0.1
asyncio==3.4.3
TgCrypto==1.2.5
